Simplify project save/autosave logic
The code for project saving was originally written for minimal
functionality, but since then has been laden with enhancements,
without any refactoring. This is an attempt to abstract out some
of the logic into a dedicated class. More work might need to be
done later, since its still nowhere close to an ideal state.